Additional methods to discover speakers.
In addition to the automated discovery of speakers add the
ability to parse custom parameters with speaker name/ip address.

Also add the ability for the automated discovery to scan specific
subnets as the original discovery would only scan the submet that
the Polisy/eisy is on.

This also fixes a bug in the poll function so that the status
of the speakers is queried and updated at the short poll interval.